requirement.txt
absl-py==0.11.0
astor==0.8.1
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.4.5.1
chardet==3.0.4
console-progressbar==1.1.2
cycler==0.10.0
decorator==4.4.2
efficientnet==1.1.1
gast==0.2.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
icc-rt==2020.0.133
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
intel-numpy==1.15.1
intel-openmp==2020.0.133
joblib==0.17.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.3.3
matplotlib==3.2.1
mkl==2019.0
mkl-fft==1.0.6
mkl-random==1.0.1
networkx==2.5
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
pandas==1.0.4
Pillow==6.0.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
tbb==2019.0
tbb4py==2019.0
tensorboard==1.15.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.4
termcolor==1.1.0
Theano==1.0.5
threadpoolctl==2.1.0
tifffile==2020.9.3
tqdm==4.51.0
urllib3==1.26.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
